fakegitserver: code reorg
Move out code from main.go to fakegitserver.go to separate
housekeeping (boilerplate) code from business logic.

Also rename `lib` folder to `internal`, per
https://github.com/golang-standards/project-layout.

No functional change.
